seeding report database

+++ ECom.Services.Reports/Models/DailyReport.cs
@@ -14,15 +14,15 @@ public partial class DailyReport
 
     [DefaultValue(0)]
     [Column("income")]
-    public int Income { get; set; }
+    public long Income { get; set; }
 
     [DefaultValue(0)]
     [Column("outcome")]
-    public int Outcome { get; set; }
+    public long Outcome { get; set; }
 
     [DefaultValue(0)]
     [Column("profit")]
-    public int Profit { get; set; }
+    public long Profit { get; set; }
 
     [DefaultValue(0)]
     [Column("sold_quantity")]
