+++ ECom.Services.Reports/Models/MonthlyReport.cs
@@ -14,15 +14,15 @@ public partial class MonthlyReport
 
     [Column("income")]
     [DefaultValue(0)]
-    public int Income { get; set; }
+    public long Income { get; set; }
 
     [Column("outcome")]
     [DefaultValue(0)]
-    public int Outcome { get; set; }
+    public long Outcome { get; set; }
 
     [Column("profit")]
     [DefaultValue(0)]
-    public int Profit { get; set; }
+    public long Profit { get; set; }
 
     [Column("sold_quantity")]
     [DefaultValue(0)]
